nsis_test_files/requirements.txt
asgiref==3.7.2
backports.ssl_match_hostname==3.7.0.1
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
Django==4.2.2
filelock==3.17.0
fsspec==2025.2.0
idna==3.10
Jinja2==3.1.5
joblib==1.4.2
llvmlite==0.44.0
MarkupSafe==3.0.2
more-itertools==10.6.0
mpmath==1.3.0
msgpack-python==0.5.6
msgpack-rpc-python==0.4.1
networkx==3.4.2
numba==0.61.0
numpy==2.1.3
openai-whisper @ git+https://github.com/openai/whisper.git@517a43ecd132a2089d85f4ebc044728a71d49f6e
pywin32==308
regex==2024.11.6
requests==2.32.3
scikit-learn==1.6.1
scipy==1.15.1
setuptools==75.8.0
sqlparse==0.4.4
sympy==1.13.1
threadpoolctl==3.5.0
tiktoken==0.8.0
torch==2.6.0
tornado==4.5.3
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2023.3
urllib3==2.3.0
watchdog==6.0.0
